[#499] Decouple some stuff for tests
