+++ robot/rospackages/src/mcu_control/scripts/CommsNode.py
@@ -16,6 +16,7 @@ from robot.rospackages.src.mcu_control.srv import *
 
 from robot.rospackages.src.mcu_control.scripts.ArmCommands import arm_out_commands, arm_in_commands
 from robot.rospackages.src.mcu_control.scripts.RoverCommands import rover_out_commands, rover_in_commands
+from robot.rospackages.src.mcu_control.scripts.CommsUtils import parse_command, get_arg_bytes
 
 import robot.rospackages.src.mcu_control.scripts.CommsDataTypes as dt
 
@@ -129,26 +130,6 @@ def arm_command_callback(message):
 
     temp_list = [command, args, ARM_SELECTED]
     arm_queue.append(temp_list)
-
-
-def parse_command(message):
-    full_command = message.split(" ")
-    if full_command is not None:
-        command = full_command[1]
-        args = full_command[2:]
-        newArgs = []
-        for arg in args:
-            try:
-                newArgs.append(float(arg))
-            except ValueError:
-                newArgs.append(arg)
-
-        return command, newArgs
-    return None, []
-
-
-def get_arg_bytes(command_tuple):
-    return sum(element[1] for element in command_tuple[2])
 
 
 def main():
